Timelines

+++ pupil_src/shared_modules/system_graphs.py
@@ -100,7 +100,7 @@ class System_Graphs(System_Plugin_Base):
 
     def on_window_resize(self, window, *args):
         fb_size = glfw.glfwGetFramebufferSize(window)
-        hdpi_factor = float(fb_size[0] / glfw.glfwGetWindowSize(window)[0])
+        hdpi_factor = fb_size[0] / glfw.glfwGetWindowSize(window)[0]
 
         self.cpu_graph.scale = hdpi_factor
         self.fps_graph.scale = hdpi_factor
